[rrc] split the stage chain and emission knobs into driver::stage
Stage, Lto, RuntimeLinkage, SanitizerCli, and VariantEncoding — the
value-enums that define where the pipeline is entered and left and how
codegen is flavored — move to their own module.
